Fix bug with removing cities
If a user wanted to remove a city from the list without selecting one,
it would throw an error. Now, it reminds the user that they have not
selected a city yet, and bails out of the remove city function.

+++ src/index.js
@@ -71,6 +71,10 @@ function makeWeatherPage(event) {
 DOMinstance.citiesList.addEventListener('click', makeWeatherPage);
 
 function removeCity(){
+  if (DOMinstance.weatherPageCity === undefined){
+    alert('You have to select a city before you can remove it from the list.');
+    return
+  }
   const cityId = DOMinstance.weatherPageCity.id;
   database.removeCity(cityId);
   DOMinstance.removeCity(cityId);
